refactor: move warn method up
