+++ src/user-event/press/press.ts
@@ -1,15 +1,13 @@
 import { ReactTestInstance } from 'react-test-renderer';
-import { EventBuilder } from '../event-builder';
-import { UserEventConfig, UserEventInstance } from '../setup';
-import { dispatchEvent, wait } from '../utils';
 import act from '../../act';
 import { getHostParent } from '../../helpers/component-tree';
 import { filterNodeByType } from '../../helpers/filterNodeByType';
 import { isPointerEventEnabled } from '../../helpers/pointer-events';
 import { getHostComponentNames } from '../../helpers/host-component-names';
-import { jestFakeTimersAreEnabled } from '../../helpers/timers';
+import { EventBuilder } from '../event-builder';
+import { UserEventConfig, UserEventInstance } from '../setup';
+import { dispatchEvent, wait, warnAboutRealTimersIfNeeded } from '../utils';
 import { DEFAULT_MIN_PRESS_DURATION } from './constants';
-import { warnAboutRealTimers } from './utils/warnAboutRealTimers';
 
 export type PressOptions = {
   duration: number;
@@ -31,20 +29,20 @@ export async function longPress(
 }
 
 const basePress = async (
-  config: UserEventInstance['config'],
+  config: UserEventConfig,
   element: ReactTestInstance,
   options: PressOptions = { duration: 0 }
 ): Promise<void> => {
   // Text and TextInput components are mocked in React Native preset so the mock
   // doesn't implement the pressability class
   // Thus we need to call the props directly on the host component
-  if (isEnabledHostText(element) || isEnabledTextInput(element)) {
-    await dispatchBasicPressEventSequence(config, element, options);
+  if (isPressableText(element) || isEnabledTextInput(element)) {
+    await emitBasicPressEvents(config, element, options);
     return;
   }
 
   if (isEnabledTouchResponder(element)) {
-    await dispatchPressablePressEventSequence(config, element, options);
+    await emitPressablePressEvents(config, element, options);
     return;
   }
 
@@ -56,15 +54,12 @@ const basePress = async (
   await basePress(config, hostParentElement, options);
 };
 
-const dispatchPressablePressEventSequence = async (
+const emitPressablePressEvents = async (
   config: UserEventConfig,
   element: ReactTestInstance,
   options: PressOptions = { duration: 0 }
 ) => {
-  const areFakeTimersEnabled = jestFakeTimersAreEnabled();
-  if (!areFakeTimersEnabled) {
-    warnAboutRealTimers();
-  }
+  warnAboutRealTimersIfNeeded();
 
   await wait(config);
 
@@ -83,6 +78,9 @@ const dispatchPressablePressEventSequence = async (
       EventBuilder.Common.responderRelease()
     );
 
+    // React Native will wait for minimal delay of DEFAULT_MIN_PRESS_DURATION
+    // before emitting the `pressOut` event. We need to wait here, so that
+    // `press()` function does not return before that.
     if (DEFAULT_MIN_PRESS_DURATION - options.duration > 0) {
       await wait(config, DEFAULT_MIN_PRESS_DURATION - options.duration);
     }
@@ -96,7 +94,7 @@ const isEnabledTouchResponder = (element: ReactTestInstance) => {
   );
 };
 
-const isEnabledHostText = (element: ReactTestInstance) => {
+const isPressableText = (element: ReactTestInstance) => {
   return (
     filterNodeByType(element, getHostComponentNames().text) &&
     isPointerEventEnabled(element) &&
@@ -117,8 +115,8 @@ const isEnabledTextInput = (element: ReactTestInstance) => {
  * Dispatches a basic press event sequence on non-Pressable component,
  * e.g. Text or TextInput.
  */
-async function dispatchBasicPressEventSequence(
-  config: UserEventInstance['config'],
+async function emitBasicPressEvents(
+  config: UserEventConfig,
   element: ReactTestInstance,
   options: PressOptions = { duration: 0 }
 ) {
