+++ src/user-event/utils/index.ts
@@ -2,3 +2,4 @@ export * from './content-size';
 export * from './dispatch-event';
 export * from './text-range';
 export * from './wait';
+export * from './warn-about-real-timers';
