+++ src/user-event/utils/warn-about-real-timers.ts
@@ -1,4 +1,11 @@
-export const warnAboutRealTimers = () => {
+import { jestFakeTimersAreEnabled } from '../../helpers/timers';
+
+export const warnAboutRealTimersIfNeeded = () => {
+  const areFakeTimersEnabled = jestFakeTimersAreEnabled();
+  if (areFakeTimersEnabled) {
+    return;
+  }
+
   // eslint-disable-next-line no-console
   console.warn(`It is recommended to use userEvent with fake timers
 Some events involve duration so your tests may take a long time to run.
